Consolidate notecard utils
